--- formula-coroutines/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=formula-coroutines
-POM_NAME=Formula Kotlin Coroutines Runtime
-POM_PACKAGING=jar
