+++ formula/build.gradle.kts
@@ -13,9 +13,9 @@ dependencies {
 
     testImplementation(project(":formula-test"))
     testImplementation(project(":formula-rxjava3"))
-    testImplementation(project(":formula-coroutines"))
     testImplementation(libs.truth)
     testImplementation(libs.junit)
     testImplementation(libs.rxrelays)
     testImplementation(libs.coroutines.test)
+    testImplementation(libs.turbine)
 }
