+++ jacoco.sh
@@ -2,7 +2,6 @@
 # Should match with what we run in .github/workflows/build-workflow.yml
 ./gradlew clean
 ./gradlew :formula:test
-./gradlew :formula-coroutines:test
 ./gradlew :formula-android:testRelease
 ./gradlew :formula-android-tests:testRelease
 ./gradlew :formula-test:test
